Initial Backoff Attempt
Attempts fine grain locking with a back off policy.

Needs to T&T&S the state passed to episode, then when selecting the
next state, it randomly selects one, looks if it is locked, if so it
tries to find a random one again. Then tries a T&S.

Before the recursive call, both states are unlocked.